TAN-1799 Clean up jsonString types
